Data Processing 6/17
This is my progress for today on the REU data processing. I've been able to read in the csv file, get rid of the unneeded columns, put the data in a (n x N) matrix, compute a measure to compare our results to (category totals), and plot the histograms. I am currently working on rescaling the data.